@@ -12,7 +12,7 @@ authors:
 homepage: https://github.com/Workiva/w_flux
 
 environment:
-    sdk: '>=2.4.0 <3.0.0'
+    sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
     meta: ^1.7.0
